pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jax-aht"
version = "1.0.0"
description = "JAX-based benchmark for Ad Hoc Teamwork research"
authors = [{name = "Learning Agents Research Group"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["reinforcement-learning", "multi-agent", "jax", "ad-hoc-teamwork", "benchmark"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # JAX AI stack v2025.2.5
    "jax[cuda12]==0.5.3",
    "chex==0.1.88",
    "flax==0.10.2",
    "ml_dtypes==0.4.0",
    "optax==0.2.4",
    "orbax-checkpoint==0.11.1",
    "orbax-export==0.0.6",
    # Core dependencies
    "numpy",
    "gymnax",
    "distrax",
    "safetensors",
    "moviepy",
    "ffmpeg",
    "prettytable",
    "rliable",
    "matplotlib",
    "wandb",
    "omegaconf",
    "hydra-core",
    "huggingface-hub",
    # Pinned to avoid arch/pandas/setuptools compatibility issues
    "setuptools<71",
    "arch>=8.0",
    "pandas<3.0",
    # Environment dependencies
    "jumanji==1.1.0",
    "jaxmarl==0.0.7",
]

[project.urls]
Homepage = "https://github.com/LARG/jax-aht"
Repository = "https://github.com/LARG/jax-aht"
Documentation = "https://github.com/LARG/jax-aht/blob/main/README.md"
"Bug Tracker" = "https://github.com/LARG/jax-aht/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "common*", "envs*", "evaluation*", "ego_agent_training*", "marl*", "open_ended_training*", "teammate_generation*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json", "*.md"]
